refacto controller+ fix front + fallback image laverie
