+++ backend/src/Repository/ProfessionnelRepository.php
@@ -3,6 +3,7 @@
 namespace App\Repository;
 
 use App\Entity\Professionnel;
+use App\Entity\Utilisateur;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 
@@ -14,6 +15,16 @@ class ProfessionnelRepository extends ServiceEntityRepository
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, Professionnel::class);
+    }
+
+    public function findOneByUtilisateur(Utilisateur $utilisateur): ?Professionnel
+    {
+        return $this->findOneBy(['utilisateur' => $utilisateur]);
+    }
+
+    public function findOneBySiren(int $siren): ?Professionnel
+    {
+        return $this->findOneBy(['siren' => $siren]);
     }
 
     //    /**
